+++ modules/client/package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "@connext/cf-core": "3.0.1",
+    "@connext/channel-provider": "3.0.1",
     "@connext/messaging": "3.0.1",
     "@connext/store": "3.0.1",
     "@connext/types": "3.0.1",
